Adding core win32 support and a dll injector/function hooker.
This adds a win32 implementation of the fd wrapper and some win32 compat
headers/source files. With those two the impact on the rest of the code is
minimal and mainly involves #ifdef'ing out unix signal related stuff and unix
specific header.

The dll injector is a dll which is injected into the compiler process via
CreateRemoteThread and copying over a chunk of assembly to run. The injected
dll then hooks the appropriate kernel32 functions by iterating over the
function pointers in the import address table and replacing them with proxy
functions.

This commit does not include the changes to the server code in order for it to
compile on windows.