Only use cmd if actually needed.
For the moment this was done as a temporary fix to get around the stupidly low
limit cmd has on command line length (~256 bytes), which is easy to hit with
link commands. Ideally we should probably use a more sh compliant shell, but
the mingw one is quite slow to launch, has permissions trouble, and would be a
very large external dependency.